+++ CHANGELOG.md
@@ -7,9 +7,12 @@
 - Core : the "legacy" es5 umd bundle is now automatically transpiled (from ES6 to ES5) using rollup and bubble
 - Core : internal rewrite on object bound implementation, with all Renderable now fully using/relying on the me.Bounds object introduced in last version
 - Container : Containers defines an additional `enableChildBoundsUpdate` flag to enable full bounds update, including child bounds (disabled by default)
+- Entity : me.CollectableEntity is now deprecated and replaced by a more generic `me.Collectable` base object that do not extend me.Entity anymore
+- Entity : me.LevelEntity is now deprecated and replaced by a more generic `me.Trigger` objects that do not extend me.Entity anymore
 - Renderer : WebGL2 is now the default mode when using the WebGL renderer (use `preferWebGL1 = true` calling me.video.init if you need to force WebGL1)
 - Tiled : add TMX `tintcolor` parsing for tile and object layers
 - Stage : `me.Stage` constructor now accept new argument properties to specify the `onResetEvent` and `onDestroyEvent` functions
+- Stage : fix a crash with legacy API when using the deprecated me.ScreenObject object
 - TMX : fix crash when trying to add collisionType to text nodes (thanks @framp)
 - TMX : fix a regression with the Tile Layer preRender feature
 - TMX : fix tile properties import for the new Tiled JSON format (thanks @3ck0o)